+++ packages/front/src/components/shared/mini-card.tsx
@@ -27,6 +27,32 @@ const protections = {
   }
 };
 
+const protections = {
+  "account balance display": {
+    message: "Don’t let people know how much money you have!",
+  },
+  "transactions tracking": {
+    message: "No one needs to know what you have been doing on chain!",
+  },
+  "profile analysis": {
+    message:
+      "Not a good idea to have people analyzing your profile based on on chain activity!",
+  },
+  "wallet connections tracking": {
+    message:
+      "You don’t want everybody to know who you have been transacting lately!",
+  },
+
+  "portfolio tracking": {
+    message:
+      "Especially for institutions, you don’t want to have your portfolio tracked by other players!",
+  },
+  "front run": {
+    message:
+      "If you’re a big investor, it’s better to be hidden to prevent front-run attacks.",
+  },
+};
+
 const MiniCard = ({
   title,
   protection
