+++ packages/front/src/pages/index.tsx
@@ -9,11 +9,6 @@ import Fade from "react-reveal/Fade";
 import HeaderLanding from "@/components/layout/header-landing";
 import Card from "@/components/shared/card";
 import MiniCard from "@/components/shared/mini-card";
-import Zoom from "react-reveal/Zoom";
-import Fade from "react-reveal/Fade";
-import "swiper/css";
-import { useEffect, useState } from "react";
-import { ProtectionModal } from "@/components/modals/protections";
 
 const Waves = () => {
   return (
